Fix warnings/errors for icc compiler
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@100873 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: a36db4e4f9e98d4a323c08243f80085bd8e65751

+++ Online/Online/ROMon/rpc/PyRPC.cpp
@@ -721,7 +721,7 @@ Server::Server() : m_server(), m_port(-1), m_channel(0) { }
 Server::~Server() {  disconnect();   }
 
 /// Connect to server
-bool Server::connect(const string& server, int port) {
+bool Server::connect(const string& server, unsigned short port) {
   NetworkChannel::Address sin;
   struct hostent *h = ::gethostbyname(server.c_str());
   disconnect();
